seven

+++ apps/books/views.py
@@ -268,7 +268,7 @@ def send_mail_to_lender(recipient_email, borrower_email, book_title):
     mail.send(fail_silently=False)    
 
 
-@api_view(["PATCH"])
+@api_view(["DELETE"])
 @permission_classes([IsActiveUser])
 @authentication_classes([JWTAuthentication])
 def cancel_borrow_request(request, request_id):
@@ -277,7 +277,7 @@ def cancel_borrow_request(request, request_id):
         return Response({"message":"This request has already been processed."}, status=status.HTTP_400_BAD_REQUEST)
     
     borrow_request.status = 'CANCELLED'
-    borrow_request.save()
+    borrow_request.delete()
     return Response({"message":"Borrow request cancelled."}, status=status.HTTP_200_OK)
 
 
@@ -312,7 +312,7 @@ def accept_borrow_request(request, request_id):
     return Response({"message":"Borrow request accepted.", "data": serializer.data}, status=status.HTTP_200_OK)
 
 
-@api_view(["POST"])
+@api_view(["DELETE"])
 @permission_classes([IsActiveUser])
 @authentication_classes([JWTAuthentication])
 def reject_borrow_request(request, request_id):
@@ -324,7 +324,7 @@ def reject_borrow_request(request, request_id):
         return Response({"message":"You are not authorized to reject this request."}, status=status.HTTP_403_FORBIDDEN)
     
     borrow_request.status = 'REJECTED'
-    borrow_request.save()
+    borrow_request.delete()
     serializer = BorrowRequestSerializer(borrow_request)
     return Response({"message":"Borrow request rejected.", "data": serializer.data}, status=status.HTTP_200_OK)       
 
@@ -334,11 +334,25 @@ def reject_borrow_request(request, request_id):
 @authentication_classes([JWTAuthentication])
 def delete_borrow_request(request, request_id):
     borrow_request = get_object_or_404(BorrowRequest, id=request_id)
+
     if borrow_request.requester != request.user and borrow_request.owner != request.user:
-        return Response({"message":"You are not authorized to delete this request."}, status=status.HTTP_403_FORBIDDEN)
-    
-    borrow_request.delete()
-    return Response({"message":"Borrow request deleted successfully."}, status=status.HTTP_200_OK)
+        return Response(
+            {"message": "You are not authorized to delete this borrow request."},
+            status=status.HTTP_403_FORBIDDEN
+        )
+
+    if borrow_request.status in ['ACCEPTED', 'RETURNED', 'PENDING']:
+        return Response(
+            {"message": "Accepted or returned requests cannot be deleted."},
+            status=status.HTTP_400_BAD_REQUEST
+        )
+    borrow_request.delete() 
+
+    return Response(
+        {"message": "Borrow request deleted successfully."},
+        status=status.HTTP_200_OK
+    )
+
 
 # borrower history
 @api_view(["GET"])
